requirements.txt
youtube-transcript-api
langchain
langchain-community
langchain-google-genai
faiss-cpu
streamlit
python-dotenv